10/26/20
Today in class I continued to work on getting my index page arranged with CSS. I hoped to finish by the end of the class, but got stuck arranging the top hero section of new albums and arranging the text in between them. I am really close but need to adjust the floats I believe. I think I am still getting used to remembering exactly what each tag does and the CSS properties that are to be used, but otherwise doing well and enjoying the process. Personal goal for myself by the end of the week is to complete the index and about pages so that way next week I can continue with tour and music then add in media queries.