10/28/20
I am finally finding my rhythm with CSS and connecting it to the design on Figma. I have again worked on CSS and updated a few minor things in the index and about pages. I think my biggest struggle/problem/puzzle is positioning correctly on the page but I know that the actual design elements are strong.