11/4/20 Updates
Tonight I worked on some details of my index page including positioning, more specific styles for text, and colors. I still need to finish, on my index page, the form in the header where the user can input their email, the contact section at the bottom, the positioning of the "zamily" instagram pictures, the positioning of the text in the "music" section, and the more specific font styles in the main hero/new album section. SO CLOSE!!!!